% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_fct.R
\name{cov.wend.interpol}
\alias{cov.wend.interpol}
\title{Calculates the Generalized Wendland covariance matrix.}
\usage{
cov.wend.interpol(h, theta, abstol = 1e-05, reltol = 0.01,
  n_interpol = 300, eps = getOption("spam.eps"))
}
\value{
If the distance matrix is in standard R format a standard R matrix is
returned. If the distance matrix is of class 'spam' the returned matrix is
also of class 'spam'.
}
\description{
The function 'cov.wend.interpol()' calculates the Generalized
Wendland (GW) covariance matrix based on a distance matrix. In
contrast to 'cov.wend()' this function uses interpolation in order to
increase calculation speed.
}
\examples{
x <- seq(0,1,len=10) 
loc <- expand.grid(x,x) 
dist.mat <- spam::nearest.dist(loc,upper=NULL,delta=0.5) 
cov.wend.interpol( dist.mat, c(0.3,6,1.5,1,0))
}
\seealso{
[spam]
}
