\name{GWcovar-package}
\alias{GWcovar-package}
\alias{GWcovar}
\docType{package}
\title{
\packageTitle{GWcovar}
}
\description{
\packageDescription{GWcovar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GWcovar}
\packageIndices{GWcovar}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GWcovar}

Maintainer: \packageMaintainer{GWcovar}
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
%\examples{
%~~ simple examples of the most important functions ~~
%}
