% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_fct.R
\docType{package}
\name{GWcovar}
\alias{GWcovar}
\alias{GWcovar-package}
\title{Calculation of the Generalized Wendland Covariance Matri}
\description{
Two functions to calculate the GW covariance matrix. For the distance
matrix eigher sparse matrices from the package \linkS4class{spam} or
standard R matrices can be used.  The package provides the two
functions \code{\link{cov.wend}} and \code{\link{cov.wend.interpol}}.
}
\references{
\url{https://arxiv.org/abs/1607.06921}
}
\seealso{
\pkg{spam}
}
\author{
Josef Stocker, \email{josef@josefstocker.ch}
}
