% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_fct.R
\docType{package}
\name{GWcovar}
\alias{GWcovar}
\alias{GWcovar-package}
\title{GWcovar provides functions to calculate the Generalized Wendland
covariance matrix}
\description{
The package GWcovar provides the two functions
\code{\link{cov.wend()}} and \code{\link{cov.wend.interpol()}}
}
