% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_fct.R
\docType{package}
\name{GWcovar}
\alias{GWcovar}
\alias{GWcovar-package}
\title{This package provides functions to calculate the Generalized Wendland
covariance matrix}
\description{
The package rovides the two functions
\code{\link{cov.wend}} and \code{\link{cov.wend.interpol}}.
}
\references{
\url{https://arxiv.org/abs/1607.06921}
}
\seealso{
\pkg{spam}
}
\author{
Josef Stocker, \email{josef@josefstocker.ch}
}
